Disable iterationDelay delay on the beginning of animation (#324)
In the current version, when an "iterationDelay" is applied to pause the animation between iterations, the same iterationDelay is applied on the beginning of the first animation too. To apply pause/delay before the beginning of first animation, there already exists the "delay" prop.

Changes in this pull request will make sure the"iterationDelay" is only applied to the second and subsequent animations.